Refactor user schema to record/api schema
