+++ backend/compact-connect/lambdas/python/common/cc_common/data_model/user_client.py
@@ -5,7 +5,11 @@ from botocore.exceptions import ClientError
 
 from cc_common.config import _Config, logger
 from cc_common.data_model.query_paginator import paginated_query
-from cc_common.data_model.schema.user import CompactPermissionsRecordSchema, UserAttributesSchema, UserRecordSchema
+from cc_common.data_model.schema.user.record import (
+    CompactPermissionsRecordSchema,
+    UserAttributesRecordSchema,
+    UserRecordSchema,
+)
 from cc_common.exceptions import CCInvalidRequestException, CCNotFoundException
 from cc_common.utils import get_sub_from_user_attributes
 
@@ -16,7 +20,7 @@ class UserClient:
     def __init__(self, config: _Config):
         self.config = config
         self.schema = UserRecordSchema()
-        self.user_attributes_schema = UserAttributesSchema()
+        self.user_attributes_schema = UserAttributesRecordSchema()
         self.compact_permissions_schema = CompactPermissionsRecordSchema()
 
     @paginated_query
