+++ backend/compact-connect/lambdas/python/staff-users/handlers/__init__.py
@@ -1,5 +1,5 @@
 from aws_lambda_powertools import Logger
-from cc_common.data_model.schema.user import UserAPISchema
+from cc_common.data_model.schema.user.api import UserAPISchema
 
 logger = Logger()
 user_api_schema = UserAPISchema()
