+++ backend/compact-connect/lambdas/python/staff-users/tests/function/__init__.py
@@ -81,7 +81,7 @@ class TstFunction(TstLambdas):
 
     def _create_compact_staff_user(self, compacts: list[str]):
         """Create a compact-staff style user for each jurisdiction in the provided compact."""
-        from cc_common.data_model.schema.user import UserRecordSchema
+        from cc_common.data_model.schema.user.record import UserRecordSchema
 
         schema = UserRecordSchema()
 
@@ -107,7 +107,7 @@ class TstFunction(TstLambdas):
 
     def _create_board_staff_users(self, compacts: list[str]):
         """Create a board-staff style user for each jurisdiction in the provided compact."""
-        from cc_common.data_model.schema.user import UserRecordSchema
+        from cc_common.data_model.schema.user.record import UserRecordSchema
 
         schema = UserRecordSchema()
