+++ backend/compact-connect/lambdas/python/staff-users/tests/unit/test_data_model/test_schema/test_user.py
@@ -8,7 +8,7 @@ from tests import TstLambdas
 
 class TestUserRecordSchema(TstLambdas):
     def test_transform_api_to_dynamo_permissions(self):
-        from cc_common.data_model.schema.user import UserAPISchema
+        from cc_common.data_model.schema.user.api import UserAPISchema
 
         with open('tests/resources/api/user-post.json') as f:
             api_user = json.load(f)
@@ -25,7 +25,8 @@ class TestUserRecordSchema(TstLambdas):
         self.assertEqual(dynamo_user['permissions'], dumped_user['permissions'])
 
     def test_transform_dynamo_to_api_permissions(self):
-        from cc_common.data_model.schema.user import UserAPISchema, UserRecordSchema
+        from cc_common.data_model.schema.user.api import UserAPISchema
+        from cc_common.data_model.schema.user.record import UserRecordSchema
 
         with open('tests/resources/api/user-post.json') as f:
             api_user = json.load(f)
@@ -43,7 +44,7 @@ class TestUserRecordSchema(TstLambdas):
 
     def test_serde_record(self):
         """Test round-trip serialization/deserialization of user records"""
-        from cc_common.data_model.schema.user import UserRecordSchema
+        from cc_common.data_model.schema.user.record import UserRecordSchema
 
         with open('../common/tests/resources/dynamo/user.json') as f:
             expected_user = TypeDeserializer().deserialize({'M': json.load(f)})
@@ -58,7 +59,7 @@ class TestUserRecordSchema(TstLambdas):
         self.assertEqual(expected_user, user_data)
 
     def test_invalid_record(self):
-        from cc_common.data_model.schema.user import UserRecordSchema
+        from cc_common.data_model.schema.user.record import UserRecordSchema
 
         with open('../common/tests/resources/dynamo/user.json') as f:
             user_data = TypeDeserializer().deserialize({'M': json.load(f)})
